Feat: ✨ 과외 일정 기본 CRUD API 구현(#S10P12A807-144)

+++ server/src/main/java/com/ioi/haryeom/tutoring/dto/TutoringScheduleIdsResponse.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 @Getter
 public class TutoringScheduleIdsResponse {
 
-    private List<Integer> tutoringScheduleIds;
+    private List<Long> tutoringScheduleIds;
 
-    public TutoringScheduleIdsResponse(List<Integer> tutoringScheduleIds) {
+    public TutoringScheduleIdsResponse(List<Long> tutoringScheduleIds) {
         this.tutoringScheduleIds = tutoringScheduleIds;
     }
 }
